+++ Editor/Scripts/PropertyDrawers/ConditionalDrawer.cs
@@ -10,23 +10,22 @@ namespace StatusEffects.Inspector
     public class ConditionalDrawer : PropertyDrawer
     {
         private const float _padding = 3;
-        private const float _ifSize = 15;
+        private const float _ifSize = 10;
         private const float _searchablePropertySize = 70;
-        private const float _isSize = 15;
+        private const float _isSize = 10;
         private const float _existsPropertySize = 70;
-        private const float _thenSize = 30;
-        private const float _secondsPropertySize = 20;
+        private const float _thenSize = 27;
+        private const float _secondsPropertySize = 28;
 
         SerializedProperty searchable;
         SerializedProperty exists;
         SerializedProperty add;
         SerializedProperty configurable;
         SerializedProperty duration;
-        SerializedProperty timed;
+        SerializedProperty timing;
 
         private Existence existence;
         private Configurability configurability;
-        private Timing timing;
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
@@ -35,15 +34,15 @@ namespace StatusEffects.Inspector
             add = property.FindPropertyRelative("add");
             configurable = property.FindPropertyRelative("configurable");
             duration = property.FindPropertyRelative("duration");
-            timed = property.FindPropertyRelative("timed");
+            timing = property.FindPropertyRelative("timing");
 
             float width = (position.width - _ifSize 
                                           - _searchablePropertySize 
                                           - _isSize 
                                           - _existsPropertySize 
                                           - _thenSize 
-                                          - (add.boolValue && timed.boolValue ? _secondsPropertySize : 0) 
-                                          - _padding * (add.boolValue && timed.boolValue ? 5 : 3)) 
+                                          - (add.boolValue && timing.enumValueIndex == (int)Timing.Duration ? _secondsPropertySize : 0) 
+                                          - _padding * (add.boolValue && timing.enumValueIndex == (int)Timing.Duration ? 8 : add.boolValue ? 7 : 6)) 
                                           / (add.boolValue ? 3 : 2);
 
             Rect offset = new Rect(position.position, new Vector2(width, position.height));
@@ -51,20 +50,20 @@ namespace StatusEffects.Inspector
             EditorGUI.BeginProperty(position, label, property);
 
             EditorGUI.LabelField(new Rect(offset.position, new Vector2(_ifSize, offset.height)), "If");
-            offset.x += _ifSize;
+            offset.x += _ifSize + _padding;
 
             EditorGUI.PropertyField(new Rect(offset.position, new Vector2(_searchablePropertySize, offset.height)), searchable, GUIContent.none);
             offset.x += _searchablePropertySize + _padding;
 
             EditorGUI.LabelField(new Rect(offset.position, new Vector2(_ifSize, offset.height)), "is");
-            offset.x += _isSize;
+            offset.x += _isSize + _padding;
 
             existence = (Existence)Convert.ToInt32(exists.boolValue);
             exists.boolValue = Convert.ToBoolean(EditorGUI.EnumPopup(new Rect(offset.position, new Vector2(_existsPropertySize, offset.height)), existence));
             offset.x += _existsPropertySize + _padding;
 
             EditorGUI.LabelField(new Rect(offset.position, new Vector2(_thenSize, offset.height)), "then");
-            offset.x += _thenSize;
+            offset.x += _thenSize + _padding;
 
             configurability = (Configurability)Convert.ToInt32(add.boolValue);
             add.boolValue = Convert.ToBoolean(EditorGUI.EnumPopup(offset, configurability));
@@ -76,14 +75,13 @@ namespace StatusEffects.Inspector
             {
                 offset.x += width + _padding;
 
-                if (timed.boolValue)
+                if (timing.enumValueIndex == (int)Timing.Duration)
                 {
                     EditorGUI.PropertyField(new Rect(offset.position, new Vector2(_secondsPropertySize, offset.height)), duration, GUIContent.none);
                     offset.x += _secondsPropertySize + _padding;
                 }
 
-                timing = (Timing)Convert.ToInt32(timed.boolValue);
-                timed.boolValue = Convert.ToBoolean(EditorGUI.EnumPopup(offset, timing));
+                EditorGUI.PropertyField(offset, timing, GUIContent.none);
             }
 
             EditorGUI.EndProperty();
@@ -99,11 +97,6 @@ namespace StatusEffects.Inspector
         {
             Remove,
             Add
-        }
-        private enum Timing
-        {
-            Infinite,
-            Seconds
         }
     }
 }